pyproject.toml
[tool.poetry]
name = "churchsuite-chat"
version = "0.1.0"
description = "AI chatbot integrated with ChurchSuite"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.14"
httpx = "^0.25.2"
openai = "^1.3.0"
qdrant-client = "^1.14.2"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
typing-extensions = "^4.9.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
